+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Reinstate simplified `AbstractProcessorTest.outputFilesIfEnabled(Compilation, Filter)` method.
+
 ### [v0.57](https://github.com/realityforge/proton/tree/v0.57) (2022-04-27) · [Full Changelog](https://github.com/spritz/spritz/compare/v0.56...v0.57)
 
 Changes in this release: